+++ src/main/scala/io/frontroute/Directive.scala
@@ -1,16 +1,15 @@
 package io.frontroute
 
-import app.tulz.tuplez._
 import com.raquo.airstream.eventstream.EventStream
 import com.raquo.airstream.signal.Signal
 import com.raquo.airstream.signal.Var
 
 class Directive[L](
   val tapply: (L => Route) => Route
-)(implicit val ev: Tuple[L]) {
+) {
   self =>
 
-  def tflatMap[R: Tuple](next: L => Directive[R]): Directive[R] = {
+  def flatMap[R](next: L => Directive[R]): Directive[R] = {
     Directive[R] { inner =>
       self.tapply { value => (location, previous, state) =>
         next(value).tapply(inner)(location, previous, state.path(".flatMap"))
@@ -18,7 +17,7 @@ class Directive[L](
     }
   }
 
-  def tmap[R: Tuple](f: L => R): Directive[R] =
+  def map[R](f: L => R): Directive[R] =
     Directive[R] { inner =>
       self.tapply { value => (location, previous, state) =>
         inner(f(value))(location, previous, state.path(".map"))
@@ -34,7 +33,7 @@ class Directive[L](
           inner(value)(ctx, previous, state.leaveDisjunction())
         }(ctx, previous, state.enterDisjunction())
         .flatMap {
-          case complete: RouteResult.Complete => EventStream.fromValue(complete, emitOnce = true)
+          case complete: RouteResult.Complete => EventStream.fromValue(complete, emitOnce = false)
           case RouteResult.Rejected =>
             other.tapply { value => (ctx, previous, state) =>
               inner(value)(ctx, previous, state.leaveDisjunction())
@@ -44,7 +43,7 @@ class Directive[L](
 
   }
 
-  def tcollect[R: Tuple](f: PartialFunction[L, R]): Directive[R] =
+  def collect[R](f: PartialFunction[L, R]): Directive[R] =
     Directive[R] { inner =>
       self.tapply { value => (location, previous, state) =>
         if (f.isDefinedAt(value)) {
@@ -55,7 +54,7 @@ class Directive[L](
       }
     }
 
-  def tfilter(predicate: L => Boolean): Directive[L] =
+  def filter(predicate: L => Boolean): Directive[L] =
     Directive[L] { inner =>
       self.tapply { value => (location, previous, state) =>
         if (predicate(value)) {
@@ -66,11 +65,28 @@ class Directive[L](
       }
     }
 
+  def signal: Directive[Signal[L]] =
+    new Directive[Signal[L]]({ inner => (ctx, previous, state) =>
+      this.tapply {
+        value => // TODO figure this out, when this is run, enter is not yet called
+          (ctx, previous, state) =>
+            val next = state.unsetValue().path(".signal")
+            previous.getValue[Var[L]](next.path) match {
+              case None =>
+                val var$ = Var(value)
+                inner(var$.signal)(ctx, previous, next.setValue(var$))
+              case Some(var$) =>
+                var$.writer.onNext(value)
+                inner(var$.signal)(ctx, previous, next.setValue(var$))
+            }
+      }(ctx, previous, state)
+    })
+
 }
 
 object Directive {
 
-  def apply[L: Tuple](f: (L => Route) => Route): Directive[L] = {
+  def apply[L](f: (L => Route) => Route): Directive[L] = {
     new Directive[L](inner =>
       (ctx, previous, state) =>
         f(value =>
@@ -79,44 +95,6 @@ object Directive {
           }
         )(ctx, previous, state)
     )
-  }
-
-  implicit def toDirective[L: Tuple](route: Route): Directive[L] =
-    Directive[L](_ => route)
-
-  implicit class SingleValueModifiers[L](underlying: Directive1[L]) extends AnyRef {
-
-    def map[R](f: L => R): Directive1[R] =
-      underlying.tmap { case Tuple1(value) => Tuple1(f(value)) }
-
-    def flatMap[R](f: L => Directive1[R]): Directive1[R] =
-      underlying.tflatMap { case Tuple1(value) => f(value) }
-
-    def collect[R: Tuple](f: PartialFunction[L, R]): Directive[R] =
-      underlying.tcollect { case Tuple1(value) =>
-        f(value)
-      }
-
-    def filter(predicate: L => Boolean): Directive1[L] =
-      underlying.tfilter { case Tuple1(value) => predicate(value) }
-
-    def signal: Directive1[Signal[L]] =
-      new Directive[Tuple1[Signal[L]]]({ inner => (ctx, previous, state) =>
-        underlying.tapply {
-          value => // TODO figure this out, when this is run, enter is not yet called
-            (ctx, previous, state) =>
-              val next = state.unsetValue().path(".signal")
-              previous.getValue[Var[L]](next.path) match {
-                case None =>
-                  val var$ = Var(value._1)
-                  inner(Tuple1(var$.signal))(ctx, previous, next.setValue(var$))
-                case Some(var$) =>
-                  var$.writer.onNext(value._1)
-                  inner(Tuple1(var$.signal))(ctx, previous, next.setValue(var$))
-              }
-        }(ctx, previous, state)
-      })
-
   }
 
 }