+++ src/main/scala/io/frontroute/LinkHandler.scala
@@ -1,46 +1,68 @@
 package io.frontroute
 
 import org.scalajs.dom
+import org.scalajs.dom.ext._
 import org.scalajs.dom.raw._
 
 import scala.scalajs.js
 import scala.scalajs.js.annotation.JSGlobal
 
 object LinkHandler {
+
   @js.native
   @JSGlobal("window")
-  private object GlobalWindow extends js.Object {
-    var routeTo: js.Function1[String, Unit] = js.native
+  private object WindowWithRouteTo extends js.Object {
+    var routeTo: js.UndefOr[js.Function1[String, Unit]] = js.native
   }
 
-  def install(): Unit = {
-    GlobalWindow.routeTo = path => BrowserNavigation.pushState(null, null, path)
-    dom.document.body.onclick = e => {
-      findParent("a", e.target.asInstanceOf[Node]).fold(true) { element =>
-        val anchor = element.asInstanceOf[HTMLAnchorElement]
-        val rel    = anchor.rel
-        if (js.isUndefined(rel) || rel == null || rel == "") {
-          BrowserNavigation.pushState(url = anchor.href)
-          false
-        } else {
-          rel match {
-            case "external" =>
-              e.preventDefault()
-              dom.window.open(anchor.href)
-              false
-
-            case _ =>
-              true
-          }
+  private val clickListener: js.Function1[Event, Boolean] = event => {
+    findParent("a", event.currentTarget.asInstanceOf[Node]).fold(true) { element =>
+      val anchor = element.asInstanceOf[HTMLAnchorElement]
+      val rel    = anchor.rel
+      if (js.isUndefined(rel) || rel == null || rel == "") {
+        BrowserNavigation.pushState(url = anchor.href)
+        false
+      } else {
+        rel match {
+          case "external" =>
+            event.preventDefault()
+            dom.window.open(anchor.href)
+            false
+          case _ =>
+            true
         }
       }
     }
   }
 
+  private val observer = new MutationObserver((records, _) => {
+    records.foreach { record =>
+      record.addedNodes.foreach { node =>
+        node.addEventListener("click", clickListener)
+      }
+      record.removedNodes.foreach { node =>
+        node.removeEventListener("click", clickListener)
+      }
+    }
+  })
+
+  def uninstall(): Unit = {
+    observer.disconnect()
+    WindowWithRouteTo.routeTo = js.undefined
+  }
+
+  private val routeTo: js.Function1[String, Unit] = (path: String) => BrowserNavigation.pushState(null, null, path)
+
+  def install(): Unit = {
+    WindowWithRouteTo.routeTo = routeTo
+    observer.observe(dom.document, js.Dynamic.literal(childList = true).asInstanceOf[MutationObserverInit])
+    dom.document.querySelectorAll("a").foreach(_.addEventListener("click", clickListener))
+  }
+
   @scala.annotation.tailrec
-  private def findParent(tagName: String, element: Node): Option[Node] = {
+  private def findParent(tagName: String, element: Node): js.UndefOr[Node] = {
     if (js.isUndefined(element) || element == null) {
-      Option.empty
+      js.undefined
     } else {
       val tagMatched =
         element.nodeName
@@ -51,7 +73,7 @@ object LinkHandler {
           .map(_.toLowerCase())
           .contains(tagName.toLowerCase())
       if (tagMatched) {
-        Some(element)
+        element
       } else {
         findParent(tagName, element.parentNode)
       }