+++ src/main/scala/io/frontroute/RouteLocation.scala
@@ -3,12 +3,17 @@ package io.frontroute
 import io.frontroute.internal.HistoryState
 
 final case class RouteLocation(
+  hostname: String,
+  port: String,
+  protocol: String,
+  host: String,
+  origin: Option[String],
   unmatchedPath: List[String],
   params: Map[String, Seq[String]],
   state: Option[HistoryState]
 ) {
 
-  def withUnmatchedPath(path: List[String]): RouteLocation = this.copy(unmatchedPath = path)
+  @inline def withUnmatchedPath(path: List[String]): RouteLocation = this.copy(unmatchedPath = path)
 
   override def toString: String =
     s"${unmatchedPath.mkString("/")}${if (params.nonEmpty) "?" else ""}${params